Codigo/requirements_Version3.txt
pymodbus==3.7.4
pyyaml==6.0.2
pyserial==3.5